Moved global data helper to settings since it's only relevant there
