+++ apps/admin-x-framework/src/test/acceptance.ts
@@ -130,13 +130,6 @@ export function toggleLabsFlag(flag: string, value: boolean) {
     }
 }
 
-export const globalDataRequests = {
-    browseSettings: {method: 'GET', path: /^\/settings\/\?group=/, response: responseFixtures.settings},
-    browseConfig: {method: 'GET', path: '/config/', response: responseFixtures.config},
-    browseSite: {method: 'GET', path: '/site/', response: responseFixtures.site},
-    browseMe: {method: 'GET', path: '/users/me/?include=roles', response: responseFixtures.me}
-};
-
 export const settingsWithStripe = updatedSettingsResponse([
     {key: 'stripe_connect_publishable_key', value: 'pk_test_123'},
     {key: 'stripe_connect_secret_key', value: 'sk_test_123'},
