+++ apps/admin-x-settings/test/acceptance/advanced/integrations/custom.test.ts
@@ -1,7 +1,8 @@
 import {Integration, IntegrationsResponseType} from '@tryghost/admin-x-framework/api/integrations';
 import {Webhook, WebhooksResponseType} from '@tryghost/admin-x-framework/api/webhooks';
-import {chooseOptionInSelect, globalDataRequests, limitRequests, mockApi, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
+import {chooseOptionInSelect, limitRequests, mockApi, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
 import {expect, test} from '@playwright/test';
+import {globalDataRequests} from '../../../utils/acceptance';
 
 test.describe('Custom integrations', async () => {
     test('Supports creating an integration and adding webhooks', async ({page}) => {
