+++ apps/admin-x-settings/test/acceptance/advanced/integrations/firstPromoter.test.ts
@@ -1,5 +1,6 @@
-import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi, responseFixtures, updatedSettingsResponse} from '@tryghost/admin-x-framework/test/acceptance';
+import { expect, test } from '@playwright/test';
+import { mockApi, responseFixtures, updatedSettingsResponse } from '@tryghost/admin-x-framework/test/acceptance';
+import { globalDataRequests } from '../../../utils/acceptance';
 
 test.describe('First Promoter integration', async () => {
     test('Supports toggling and filling in First Promoter integration', async ({page}) => {
