+++ apps/admin-x-settings/test/acceptance/advanced/integrations/zapier.test.ts
@@ -1,6 +1,7 @@
 import {Integration, IntegrationsResponseType} from '@tryghost/admin-x-framework/api/integrations';
 import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
+import {globalDataRequests} from '../../../utils/acceptance';
+import {mockApi, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
 
 test.describe('Zapier integration settings', async () => {
     test('Showing and regenerating API keys', async ({page}) => {
