+++ apps/admin-x-settings/test/acceptance/layout.test.ts
@@ -1,5 +1,6 @@
 import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi} from '@tryghost/admin-x-framework/test/acceptance';
+import {globalDataRequests} from '../utils/acceptance';
+import {mockApi} from '@tryghost/admin-x-framework/test/acceptance';
 
 test.describe('Layout', async () => {
     test('Confirms when leaving if a section is dirty', async ({page}) => {
