+++ apps/admin-x-settings/test/acceptance/routing.test.ts
@@ -1,5 +1,6 @@
-import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi} from '@tryghost/admin-x-framework/test/acceptance';
+import { expect, test } from '@playwright/test';
+import { mockApi } from '@tryghost/admin-x-framework/test/acceptance';
+import { globalDataRequests } from '../utils/acceptance';
 
 test.describe('Routing', async () => {
     test('Reopens the opened modal when refreshing the page', async ({page}) => {
