+++ apps/admin-x-settings/test/acceptance/site/design.test.ts
@@ -1,5 +1,6 @@
-import {chooseOptionInSelect, globalDataRequests, mockApi, mockSitePreview, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
+import {chooseOptionInSelect, mockApi, mockSitePreview, responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
 import {expect, test} from '@playwright/test';
+import {globalDataRequests} from '../../utils/acceptance';
 
 test.describe('Design settings', async () => {
     test('Working with the preview', async ({page}) => {
